Add imaging control server support for framegrabber
Passing the server API endpoint to the --imaging-control flag tells the
script to take the list of framegrabber devices from heartbeat messages sent
by the imaging control server over a websocket.

To minimize (not eliminate) the potential for race conditions, frames are
captured immediately but are not submitted until the following heartbeat,
which gives us the most up-to-date camera name for that framegrabber.